Remove AgentExtraClassLoader
